feat: add support for tools, franchises, languages, game modes, and perspectives - Added `Tool`, `Franchise`, and `Language` models with migrations and traits - Added enums for game modes and player perspectives - Updated media models to support relationships with tools, franchises, and languages - Added `time_to_beat` fields and `website_urls` to `Game` model

+++ database/migrations/2021_05_08_202339_create_songs_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
         Schema::create(Song::TABLE_NAME, function (Blueprint $table) {
             $table->id();
             $table->string('amazon_id')->unique()->nullable();
-            $table->unsignedInteger('am_id')->unique()->nullable();
+            $table->unsignedBigInteger('am_id')->unique()->nullable();
             $table->unsignedInteger('deezer_id')->unique()->nullable();
             $table->unsignedInteger('mal_id')->unique()->nullable();
             $table->string('spotify_id')->unique()->nullable();
